bot_service/requirements.txt
aiogram>=3.0,<4
httpx>=0.24,<1
openai>=1.0,<2
python-dotenv>=1.0,<2
